separate pzem and sensor entitites
