--- src/modules/pzems/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './pzem-item.entity';
-export * from './pzem.entity';
