+++ src/modules/pzems/index.ts
@@ -1,8 +1,3 @@
-export * from './entities';
-export * from './pzems.constants';
 export * from './pzems.controller';
-export * from './pzems.gateway';
 export * from './pzems.module';
-export * from './pzems.repository';
 export * from './pzems.service';
-export * from './pzems.types';
