+++ src/modules/pzems/pzems.module.ts
@@ -1,15 +1,11 @@
 import { Module } from '@nestjs/common';
 import { PzemsService } from './pzems.service';
 import { PzemsController } from './pzems.controller';
-import { PzemsRepository } from './pzems.repository';
 import { EspApiModule } from '@api/modules/esp';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Pzem, PzemItem } from './entities';
-import { PzemsGateway } from './pzems.gateway';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Pzem, PzemItem]), EspApiModule],
+  imports: [EspApiModule],
   controllers: [PzemsController],
-  providers: [PzemsService, PzemsRepository, PzemsGateway],
+  providers: [PzemsService],
 })
 export class PzemsModule {}
