+++ src/modules/sensors/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './sensor-item.entity';
+export * from './sensor.entity';
