+++ src/modules/sensors/index.ts
@@ -1,0 +1,7 @@
+export * from './entities';
+export * from './sensors.constants';
+export * from './sensors.gateway';
+export * from './sensors.module';
+export * from './sensors.repository';
+export * from './sensors.service';
+export * from './sensors.types';
