+++ src/modules/sensors/sensors.types.ts
